requirements.txt
python-binance
pydantic
pyyaml
rich
requests
openai
